requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
